improve drag and drop service and fix problem with door open

+++ client/src/app/services/drag-and-drop/drag-and-drop.service.ts
@@ -1,39 +1,44 @@
 import { CdkDragDrop } from '@angular/cdk/drag-drop';
 import { Injectable } from '@angular/core';
-import { Tile } from '@app/interfaces/tile.interface';
+import { Cell } from '@app/interfaces/cell.interface';
 import { GridService } from '@app/services/grid/grid.service';
 import { TileService } from '@app/services/tile/tile.service';
+import { BehaviorSubject } from 'rxjs/internal/BehaviorSubject';
 import { OBJECTS_LIST } from 'src/constants/objects-constants';
 
 @Injectable({
     providedIn: 'root',
 })
 export class DragDropService {
-    tile: Tile;
-    objectsList = OBJECTS_LIST;
-    startedPointsIndexInList = this.objectsList.findIndex((obj) => obj.name === 'Started Points');
-    randomItemsIndexInList = this.objectsList.findIndex((obj) => obj.name === 'Random Items');
+    private cell: Cell = { row: 0, col: 0, tile: '', object: '', isOccuped: false };
+    public objectsList = [...OBJECTS_LIST];
+    private startedPointsIndexInList: number;
+    private randomItemsIndexInList: number;
+    // private gridComponent: GridComponent;
+    private objectsListSubject = new BehaviorSubject(OBJECTS_LIST);
+    objectsList$ = this.objectsListSubject.asObservable();
+
+    // Méthodes pour modifier la liste
+    updateObjectList(newList: any[]): void {
+        this.objectsListSubject.next(newList);
+    }
 
     constructor(
+        //private gridComponent: GridComponent,
+
         private gridService: GridService,
         private tileService: TileService,
     ) {
-        this.tile = { x: 0, y: 0, image: [], isOccuped: false };
+        this.randomItemsIndexInList = this.objectsList.findIndex((obj) => obj.name === 'Random Items');
+        this.startedPointsIndexInList = this.objectsList.findIndex((obj) => obj.name === 'Started Points');
     }
     drop(event: CdkDragDrop<unknown[]>, index: number): void {
-        const validDropZone: boolean = this.isDropZoneValid(event.event.target as Element);
-        if (validDropZone) {
-            this.gridService.addObjectToTile(this.tile.x, this.tile.y, event.item.data);
-            this.tile.isOccuped = true;
+        const isDropZoneValid: boolean = this.isDropZoneValid(event.event.target as Element);
+        if (isDropZoneValid) {
+            this.gridService.addObjectToTile(this.cell.row, this.cell.col, event.item.data);
 
-            if (this.objectsList[index] === this.objectsList[this.randomItemsIndexInList]) {
-                if (this.counter(index)) {
-                    return;
-                }
-            }
-
-            if (this.objectsList[index] === this.objectsList[this.startedPointsIndexInList]) {
-                if (this.counter(index)) {
+            if (this.isSpecialObject(index)) {
+                if (this.decrementObjectCounter(index)) {
                     return;
                 }
             }
@@ -42,23 +47,21 @@ export class DragDropService {
         }
     }
 
+    private isSpecialObject(index: number): boolean {
+        return index === this.randomItemsIndexInList || index === this.startedPointsIndexInList;
+    }
+
     dropObjectBetweenCase(event: CdkDragDrop<{ image: string; row: number; col: number }>, element: Element): void {
+        const isDropZoneValid: boolean = this.isDropZoneValid(event.event.target as Element);
         const { row: previousRow, col: previousCol, image: objectToMove } = event.item.data;
         const { row: currentRow, col: currentCol } = event.container.data;
-        if (objectToMove) {
+        if (objectToMove && isDropZoneValid) {
             this.tileService.removeObjectFromTile(previousRow, previousCol, objectToMove);
             this.tileService.addObjectToTile(currentRow, currentCol, objectToMove);
         }
         if (element.classList.contains('drop-zone2') || objectToMove.isDragAndDrop) {
             this.tileService.removeObjectFromTile(currentRow, currentCol, objectToMove);
-            for (const object of this.objectsList) {
-                if (object.link === objectToMove) {
-                    object.isDragAndDrop = false;
-                    if (object.count !== undefined) {
-                        object.count += 1;
-                    }
-                }
-            }
+            this.incrementObjectCounter(objectToMove);
             return;
         }
     }
@@ -66,14 +69,11 @@ export class DragDropService {
     isDropZoneValid(element: Element | null): boolean {
         while (element) {
             if (element.classList.contains('drop-zone')) {
-                const x = (this.tile.x = parseInt(element.id.split(',')[0], 10));
-                const y = (this.tile.y = parseInt(element.id.split(',')[1], 10));
+                const row = (this.cell.row = parseInt(element.id.split(',')[0], 10));
+                const col = (this.cell.col = parseInt(element.id.split(',')[1], 10));
 
-                if (this.tile.image.length >= 2) {
-                    this.tile.image = [];
-                }
-                if (x >= 0 && y >= 0 && !this.gridService.gridTiles[y][x].isOccuped && !this.isDoorOrWallTile(element)) {
-                    this.tile.image.push(element.id.split(',')[2] as string);
+                if (row >= 0 && col >= 0 && !this.gridService.gridTiles[col][row].isOccuped && !this.isDoorOrWallTile(element)) {
+                    this.cell.object = element.id.split(',')[2] as string;
                     return true;
                 } else return false;
             }
@@ -83,7 +83,17 @@ export class DragDropService {
         return false;
     }
 
-    counter(index: number): boolean {
+    incrementObjectCounter(objectToMove: string): void {
+        for (const object of this.objectsList) {
+            if (object.link === objectToMove) {
+                object.isDragAndDrop = false;
+                if (object.count !== undefined) {
+                    object.count += 1;
+                }
+            }
+        }
+    }
+    decrementObjectCounter(index: number): boolean {
         const object = this.objectsList[index];
 
         if (object && typeof object.count === 'number' && object.count > 1) {
@@ -99,12 +109,12 @@ export class DragDropService {
     isDoorOrWallTile(element: Element | null): boolean {
         while (element) {
             if (element.classList.contains('drop-zone')) {
-                const x = (this.tile.x = parseInt(element.id.split(',')[0], 10));
-                const y = (this.tile.y = parseInt(element.id.split(',')[1], 10));
+                const row = (this.cell.row = parseInt(element.id.split(',')[0], 10));
+                const col = (this.cell.col = parseInt(element.id.split(',')[1], 10));
                 if (
-                    this.gridService.getGridTiles()[y][x].images[0] === 'assets/tiles/Door.png' ||
-                    this.gridService.getGridTiles()[y][x].images[0] === 'assets/tiles/Wall.png' ||
-                    this.gridService.getGridTiles()[y][x].images[0] === 'assets/tiles/DoorOpen.png'
+                    this.gridService.gridTiles[col][row].images[0] === 'assets/tiles/Door-Open.png' ||
+                    this.gridService.gridTiles[col][row].images[0] === 'assets/tiles/Door.png' ||
+                    this.gridService.gridTiles[col][row].images[0] === 'assets/tiles/Wall.png'
                 ) {
                     return true;
                 } else return false;
